install:fix: basically, coverage was totally broken
The hash map that tracks coverage uses a linked list
for hash conflicts. The space for new elements on the
list was allocated from a stack-like data structure.
The stack pointer on it was local to each process.
That's wrong because every new iteration of fuzzing
overwritten the results of the previous runs.
To fix it, I replaced the local variable with a
global variable in the shared memory.