SpecFuzz pass: Don't instrument MFENCE instructions
Before this change, the SpecFuzz pass would fail to compile code that used
MFENCEs because it is treated as a write instruction with no memory operands by
LLVM.

This patch skips instrumenting MFENCEs. In the future, it might make sense to
have SpecFuzz respond in some way to MFENCEs, but for now, skipping them is a
decent workaround.